refactor: move input function to controller, split run steps
